Fix msg dtype: cast to page_summary dtype, not float32

+++ scripts/atom_lora.py
@@ -521,12 +521,9 @@ class AtomHypernetwork(nn.Module):
                 pad = [torch.zeros_like(msg_list[0])] * (self.max_messages - len(msg_list))
                 msg_list = msg_list + pad
             msg_cat = torch.cat(msg_list[:self.max_messages], dim=-1)  # (B, message_dim * max_messages)
-            msg_summary = self.message_project(msg_cat.float())  # (B, 256)
+            msg_summary = self.message_project(msg_cat.to(dtype=page_summary.dtype))  # (B, 256)
         else:
             msg_summary = torch.zeros(batch_size, 256, device=device)
-
-        # Cast msg_summary to match page_summary dtype (bfloat16)
-        msg_summary = msg_summary.to(dtype=page_summary.dtype)
 
         # Generate contextual logits (with or without pass embedding)
         if self.skip_pass_embed:
